+++ examples/typegraphql/index.ts
@@ -3,7 +3,7 @@ import 'reflect-metadata';
 import fastify from 'fastify';
 import { envelop, useLogger, useAsyncSchema, useTiming } from '@envelop/core';
 import { Field, ObjectType, buildSchema, ID, Resolver, Query, Arg } from 'type-graphql';
-import { getGraphQLParameters, processRequest, renderGraphiQL, shouldRenderGraphiQL } from 'graphql-helix';
+import { getGraphQLParameters, processRequest, renderGraphiQL, sendResult, shouldRenderGraphiQL } from 'graphql-helix';
 
 @ObjectType()
 class Recipe {
@@ -84,14 +84,7 @@ app.route({
         contextFactory,
       });
 
-      if (result.type === 'RESPONSE') {
-        res.status(result.status);
-        res.send(result.payload);
-      } else {
-        // You can find a complete example with GraphQL Subscriptions and stream/defer here:
-        // https://github.com/contrawork/graphql-helix/blob/master/examples/fastify/server.ts
-        res.send({ errors: [{ message: 'Not Supported in this demo' }] });
-      }
+      sendResult(result, res.raw);
     }
   },
 });