+++ packages/plugins/resource-limitations/src/index.ts
@@ -6,11 +6,13 @@ import {
   FieldNode,
   GraphQLError,
   GraphQLField,
+  GraphQLInputType,
   GraphQLInt,
   GraphQLList,
   GraphQLNonNull,
   GraphQLObjectType,
   GraphQLType,
+  isScalarType,
 } from 'graphql';
 import { getArgumentValues } from 'graphql/execution/values.js';
 
@@ -21,13 +23,16 @@ const getWrappedType = (graphqlType: GraphQLType): Exclude<GraphQLType, GraphQLL
   return graphqlType;
 };
 
-const hasFieldDefConnectionArgs = (field: GraphQLField<any, any>) => {
+const isValidArgType = (type: GraphQLInputType, paginationArgumentTypes?: string[]): boolean =>
+  type === GraphQLInt || (isScalarType(type) && !!paginationArgumentTypes && paginationArgumentTypes.includes(type.name));
+
+const hasFieldDefConnectionArgs = (field: GraphQLField<any, any>, argumentTypes?: string[]) => {
   let hasFirst = false;
   let hasLast = false;
   for (const arg of field.args) {
-    if (arg.name === 'first' && arg.type === GraphQLInt) {
+    if (arg.name === 'first' && isValidArgType(arg.type, argumentTypes)) {
       hasFirst = true;
-    } else if (arg.name === 'last' && arg.type === GraphQLInt) {
+    } else if (arg.name === 'last' && isValidArgType(arg.type, argumentTypes)) {
       hasLast = true;
     } else if (hasLast && hasFirst) {
       break;
@@ -42,14 +47,26 @@ const buildMissingPaginationFieldErrorMessage = (params: { fieldName: string; ha
   (params.hasFirst && params.hasLast ? "either the 'first' or 'last'" : params.hasFirst ? "the 'first'" : "the 'last'") +
   ' field argument.';
 
-const buildInvalidPaginationRangeErrorMessage = (params: { fieldName: string; argumentName: string }) =>
+const buildInvalidPaginationRangeErrorMessage = (params: {
+  fieldName: string;
+  argumentName: string;
+  paginationArgumentMaximum: number;
+  paginationArgumentMinimum: number;
+}) =>
   `Invalid pagination argument for field ${params.fieldName}. ` +
-  `The value for the '${params.argumentName}' argument must be an integer within 1-100.`;
+  `The value for the '${params.argumentName}' argument must be an integer within ${params.paginationArgumentMinimum}-${params.paginationArgumentMaximum}.`;
 
 export const defaultNodeCostLimit = 500000;
 
+export const defaultPaginationArgumentMaximum = 100;
+
+export const defaultPaginationArgumentMinimum = 1;
+
 export type ResourceLimitationValidationRuleParams = {
   nodeCostLimit: number;
+  paginationArgumentMaximum: number;
+  paginationArgumentMinimum: number;
+  paginationArgumentTypes?: string[];
   reportNodeCost?: (cost: number, executionArgs: ExecutionArgs) => void;
 };
 
@@ -59,6 +76,7 @@ export type ResourceLimitationValidationRuleParams = {
 export const ResourceLimitationValidationRule =
   (params: ResourceLimitationValidationRuleParams): ExtendedValidationRule =>
   (context, executionArgs) => {
+    const { paginationArgumentMaximum, paginationArgumentMinimum } = params;
     const nodeCostStack: Array<number> = [];
     let totalNodeCost = 0;
 
@@ -77,7 +95,7 @@ export const ResourceLimitationValidationRule =
               let nodeCost = 1;
               connectionFieldMap.add(fieldNode);
 
-              const { hasFirst, hasLast } = hasFieldDefConnectionArgs(fieldDef);
+              const { hasFirst, hasLast } = hasFieldDefConnectionArgs(fieldDef, params.paginationArgumentTypes);
               if (hasFirst === false && hasLast === false) {
                 // eslint-disable-next-line no-console
                 console.warn('Encountered paginated field without pagination arguments.');
@@ -94,13 +112,14 @@ export const ResourceLimitationValidationRule =
                     )
                   );
                 } else if ('first' in argumentValues === true && 'last' in argumentValues === false) {
-                  // eslint-disable-next-line dot-notation
-                  if (argumentValues['first'] < 1 || argumentValues['first'] > 100) {
+                  if (argumentValues.first < paginationArgumentMinimum || argumentValues.first > paginationArgumentMaximum) {
                     context.reportError(
                       new GraphQLError(
                         buildInvalidPaginationRangeErrorMessage({
-                          fieldName: fieldDef.name,
+                          paginationArgumentMaximum,
+                          paginationArgumentMinimum,
                           argumentName: 'first',
+                          fieldName: fieldDef.name,
                         }),
                         fieldNode
                       )
@@ -110,13 +129,14 @@ export const ResourceLimitationValidationRule =
                     nodeCost = argumentValues['first'];
                   }
                 } else if ('last' in argumentValues === true && 'false' in argumentValues === false) {
-                  // eslint-disable-next-line dot-notation
-                  if (argumentValues['last'] < 1 || argumentValues['last'] > 100) {
+                  if (argumentValues.last < paginationArgumentMinimum || argumentValues.last > paginationArgumentMaximum) {
                     context.reportError(
                       new GraphQLError(
                         buildInvalidPaginationRangeErrorMessage({
-                          fieldName: fieldDef.name,
+                          paginationArgumentMaximum,
+                          paginationArgumentMinimum,
                           argumentName: 'last',
+                          fieldName: fieldDef.name,
                         }),
                         fieldNode
                       )
@@ -176,6 +196,20 @@ type UseResourceLimitationsParams = {
    */
   nodeCostLimit?: number;
   /**
+   * The custom scalar types accepted for connection arguments.
+   */
+  paginationArgumentScalars?: string[];
+  /**
+   * The maximum value accepted for connection arguments.
+   * @default 100
+   */
+  paginationArgumentMaximum?: number;
+  /**
+   * The minimum value accepted for connection arguments.
+   * @default 1
+   */
+  paginationArgumentMinimum?: number;
+  /**
    * Whether the resourceLimitations.nodeCost field should be included within the execution result extensions map.
    * @default false
    */
@@ -183,6 +217,8 @@ type UseResourceLimitationsParams = {
 };
 
 export const useResourceLimitations = (params?: UseResourceLimitationsParams): Plugin => {
+  const paginationArgumentMaximum = params?.paginationArgumentMaximum ?? defaultPaginationArgumentMaximum;
+  const paginationArgumentMinimum = params?.paginationArgumentMinimum ?? defaultPaginationArgumentMinimum;
   const nodeCostLimit = params?.nodeCostLimit ?? defaultNodeCostLimit;
   const extensions = params?.extensions ?? false;
   const nodeCostMap = new WeakMap<object, number>();
@@ -206,6 +242,9 @@ export const useResourceLimitations = (params?: UseResourceLimitationsParams): P
           rules: [
             ResourceLimitationValidationRule({
               nodeCostLimit,
+              paginationArgumentMaximum,
+              paginationArgumentMinimum,
+              paginationArgumentTypes: params?.paginationArgumentScalars,
               reportNodeCost: extensions
                 ? (nodeCost, ref) => {
                     nodeCostMap.set(ref, nodeCost);